Draft on MAB to optimize mutation probabilities
This commit creates a notebook inside the python folder that
implements a multi armed bandit (MAB) class to keep track of
rewards for choosing different mutation types. This class
is designed to wrap some function that makes choices with
probabilities.

The idea is that, by wrapping the function, we can optimize
the probabilities of each choice based on evidence we get
from previous tries.

I plan to tweak this a little bit more, double-check if
the code has bugs, do some experiments to make sure
this implementation improves the evolution, and then
start to implement it in the C++ version.