Improve mutation and add new comments in the code
One thing that was bothering me in the previous mutation that
I implemented was that the insert mutation would not be applied
if the tree already has the maximum depth --- which is too strict,
because the selected spot does not necessarily is the one with
maximum depth.

Now the mutation checks the allowed depth based the spot, not
on the tree.

This modification passed all tests.