Bug fix - insert mutation causing segmentation fault
previously, the insert mutation would use `get_op_with_arg` to find
a matching node to insert in the tree. Inside that function,
we ignore nodes that would make the final program having more
nodes than the `PARAMS['max_size']`. In rare situations
this could lead to an empty collection of matching nodes to sample
the operator, generating a segmentation fault.

This commit fixes that, by making the insert mutation be less
strict with the maximum size (just like PTC2): the program size
can exceed the maximum size by the number of the highest arity
among the operators.