Create tests to check dispatch_table behavior
It seems that the fit method is not working properly when we have
terminals of type ArrayXb. The sig_hash of the node is different
from all of the available nodes in the dispatch_table, raising an
error in dispatch_table.h:172.

This commit introduces a simple test case to reproduce the error
that I am getting. Ideally, We should fix this bug so this new
test case does not get a core dump.