Mutation returns std::optional<Program<T>>
This commit changes the behavior of variation functions to return
an `std::optional` value. Since variation is done stochastically,
and sometimes the Search Space will have an empty set of nodes to
use in the mutation, I've made this function to return an
std::nullopt if it fails to change the expression. This is also
applied when it succeeds to change the expression, but the new
expression exceeds `max_size` or `max_depth`.

I moved the check for max_size and depth for the end of the mutaiton function.
This makes the mutation work more generically and regardless of mutation type
(thus there is no need to manually write these checks for future mutation
implementations). While this could make mutations less effective (since
there is a greater chance of applying a mutation that will render an
invalid individual), i think this design is more robust and can be
integrated with the multi armed bandit learner.