Removed mutation size checks and added weight verification in mut and cx
Previously, in mutation, I was checking the size and depth of the
parent expression to avoid using mutations that would certainly
produce an offspring that exceeded the max_size or max_depth.

Some tests showed that this was decreasing overall performance, so
I decided to remove this feature.

Additionally, I made some tests and found that the random library
functions will sample and return a value even if all the weights
passed as arguments are smaller or equal to zero. To avoid this,
I added a check in cx and mutation to check if I'm not sampling
from a fully negative weight vector.