Mutation can write everything it did in `PARAMS["mutation_trace"]`
Everytime a mutation is called, if `PARAMS["write_mutation_trace"]`
is true, then a json object in `PARAMS["mutation_trace"]` will be
updated with information about what happened inside the mutation call.

This is usefull for debuging purposes, and also to get additional
information in cases where mutation failed and returned `nullopt`.

The json can be accessed in python without modifying the wrapper.