Mutation classes. bug fix in subtree mutation
This commit changes how mutation works by selecting the spot
**after** selecting the mutation. Previously, spot was being selected
without taking into account which mutation was going to happen.

Now, mutations are derivations of a baseclass that implements
at least two methods: find_spots and mutate.

find_spots will return the weights of nodes that can be selected
to apply the mutation. The idea here is to have more robust mutation,
avoiding a lot of nullopt returns.

mutate will actually change the expression, based on a given spot.

Now, the mutation function takes care of selecting the node
and performing the checks to determine if the search space holds
an alternative to apply the mutation. Oh, it also tak care of
writing the mutation trace (which I'm thinking about making it
an official feature, instead of just a debuging tool).