Updated tests to run with all the rework
At this point, the deap API still work, and all the stuff we
need to implement the island are there.

I need to implement a lot of testings, and make sure the island
work with the cpp implementations.

After that, I'm gonna do a cleaning up in the code (there's a
lot of TODOs that I'm leaving to fix). Finally, I'll be
ready to open a PR into master branch