Falling back to non-contextual bandits
Context is giving some kernel crashes and current experiments
does not support they are giving better results than other
bandits.

I am removing it from this branch and I will work on it on a
separate branch. Meanwhile I'll keep this one clean as there
are many improvements that should be merged soon