Lots of comments across files. updated eps
I was using an eps value that was too small to be represented
as float in cpp. As a consequence, this small epsilon value
that was supposed to get rid of log(0.0) was actually not working,
leading to numeric errors inside the weight optimizer for classification,
making it fail in some cases.

With this new value the optimization of classification weights is now
fully working (with all different clf metrics. previously it worked only
with log loss).

I decided to update all epsilon values in the library so they are more
meaningful and will not be rounded to just zero due to float precision