Improving archive interface and class_weights on model selection
Several improvements. This is a work in progress.
- class_weights can be user defined, unbalanced, or balanced by support.
it is easier now to define these values.
- archive is not serialized anymore, making it easier to predict from
archive. with this change, I removed hyperparameters related to enabling
archive, as it is easier to just save the entire population and archive
at the end of the run.
- removed predict_archive and predict_proba_archive, just access the
indivudlas directly. I expect that interface will be simpler now.
- bunch of new test cases. there is one that is still failing, I am
working on it.
- final model selection will also match the class weights specified,
so if the user asks for the smallest on the confidence interval, then
it should still work.