Variation is aware of locked weights. changed how I handle failed variations
I will not create something fully random if variation fails, because
the population can lose the locked nodes or weights. Instead, I try
subtree mutations and finally clone the parent if it fails.

This is how mutations are handling locked weights

Delete – should not work on nodes with locked weight
Toggle – will not work if there is a fixed weight (cannot turn on or off)
Subtree – will keep the weight
Point – will keep the weight
Insert – should “steal” the weight from the fixed node
Crossover – will keep the weight of the receiving parent