Modified complexity to ignore Mul operation from weighted nodes
That was making complexity values explore by considering intermediate
nodes when doing the recursive calculation, often leading to
overflow of the integer value.

I also updated the cpp test cases to print the min and max values
for each data type so we can manually check if the value is suitable
for the calculations we are doing.